requirements.txt
Flask
datetime
requests
numpy
statbotics
gunicorn
werkzeug.utils